Creates a HSweights object

Creates a list object with class HSweights. The river object can be obtained with river_network, weights with compute_river_weights or compute_area_weights, and grid with raster_to_HS.

create_HSweights(target, weights, source)

Arguments

target
A river network as an sf LINESTGRING object

weights
A data.frame with columns "riverID" in "zoneID", and "weight" corresponding to riverID in river, zoneID in HS and a numeric weight.

source
A HS object.

Value

Returns a HSweights object, which is a list of three components:

target: River network with routing information to be used in downscaling step.

weights: Weighting information for the downscaling step.

source: HS object with runoff timeseries information.