Computes specific runoff generated in the river segment.

Computes downscaled, river segment specific, runoff using the given HSweights object.

downscale_runoff(
  HSweights,
  pycno = FALSE,
  n = 10,
  dasy = NULL,
  verbose = FALSE
)

Arguments

HSweights
An object of class 'HSweights', obtained with compute_HSweights, or constructed with function create_HSweights.

pycno
If TRUE, ignore weights and perform pycnophylactic interpolation for every timestep. EXPERIMENTAL AND 2-3 ORDERS OF MAGNITUDE SLOWER THAN USING WEIGHTS. Default FALSE.

n
Iterations in the pycnophylactic interpolation.

dasy
name of the dasymetric variable in HSweights$weights. Experimental, used in pycnophylactic interpolation.

verbose
Print progress indication or not.

Value

The routed river network object with class HS, which has been enhanced with a runoff timeseries (list column runoff_ts. Runoff is given in \(m^3/s\).