Computed goodness-of-fit statistics for discharge predictions against observed streamflow

Computed goodness-of-fit statistics using gof for all the timeseries in the input HS object.

flow_gof(HS, verbose = FALSE)

Arguments

HS
A HS object with column discharge_ts.

verbose
Print progress indication or not.

Value

Returns a tibble with gof-statistics for all the observations in every discharge prediction in HS.