Computes and regionalizes optimal estimate from multiple discharge timeseries against observed streamflow

Algorithm routes runoff using user specified routing algorithm, for each observation point in sequence starting from the most upstream point. The optimum weight combination is regionalized to the other river segments which do not contain observation information.

optimise_region(HSrunoff, HSobs, routing = "simple", train = 0.5,
  optim_method = "CLS", bias_correction = FALSE, region_type = "upstream",
  no_station = "em", ..., verbose = FALSE)

Arguments

HSrunoff
An HSrunoff object containing runoff estimates.

HSobs
An HSobs object containing the observed timeseries for all stations.

routing
Routing algorithm to use. See accumulate_runoff for options.

train
The share of timeseries used for training period.

optim_method
Method used to optimise. Default uses constrained linear regression. See details.

bias_correction
Whether to apply bias correction. Default FALSE. If TRUE, constant bias correction is applied to all river segments. See details.

region_type
How to regionalize combination weights. See details.

no_station
How to handle river segments with no downstream stations. See details.

...
Additional parameters passed to the routing algorithm.

verbose
Whether or not print progress indicators.

Value

Returns an HSflow object with one discharge layer (the optimized streamflow), and two additional columns in HSflow$river:

OPTIMIZED_STATION: Name of the station from HSobs which' optimized weights were used at the river segment

OPTIMIZED_riverID: ID of the segment the weighted station is located in.

Details

Currently only one regionalization type is implemented: upstream uses the weights obtained for the nearest downstream station for the river segment.

Options for no_station is currently only one: "em", which stands for "ensemble mean". For those river segments with no weights assigned from a station, an ensemble mean (each discharge prediction is given equal weight) is computed.

If bias_correction is set to TRUE, constant bias correction is applied. Bias is divided equally among all river segments which are being optimised within a region. For example, if the sub-basin being optimised has 10 river segments, and bias (mean error) at the station being optimised is 10 m3/s, each river segment is added with a constant 1 m3/s runoff.