Computes weights HSweights from HSgrid and a river network The function computes weights either based on catchment areas (polygons), or line segments which intersect a polygon with runoff information. Weights are assigned so that the sum of weights for area or line features within a polygon equal to 1.

The river network needs to be a "clean", connected network. This means that connected river segments share a node at the point of confluence, and all linestrings are cut at segment intersections.

compute_HSweights(HSgrid, river, weights, aoi = NULL, basins = NULL,
  riverID = "riverID", verbose = FALSE)

Arguments

HSgrid
A 'HSgrid' object, obtained with raster_to_HSgrid.

river
An 'sf' linestring feature representing a river network.

weights
A character vector specifying type of weights, or a vector of user- specified weights. See Details.

aoi
An area of interest. 'sf' polygon object. Optional.

basins
An 'sf' polygon object. If weights are set to "area", providing basins skips the delineation process. ID column must have the name as riverID See Details. Optional.

riverID
A character string which specifies the name of the column in river containing unique river network identifiers. Defaults to "riverID".

verbose
Whether or not print progress indicators.

Value

Returns a list object with class 'HSweights', containing the following elements:

river. The supplied river network with routing information. See river_network for details.

weights. River network lines or catchment polygon network which was used as the basis of weighting. See compute_area_weights or compute_river_weights for details.

HSgrid. HSgrid object containing runoff information. See raster_to_HSgrid for details.

Details

Weights should be one of the following: "equal", "length", "strahler", "area", or a numeric vector which specifies the weight for each river segment.

equal option assigns equal weights to all river segments within a polygon.

length option weights river segments within a polygon based on the length of the segment.

strahler option weights river segments based on the Strahler number computed for the supplied river network.

A numeric vector with length equal to the number of features in river. Weights will be computed within a polygon: \(x_i/sum(x)\)

area option weights the river segments based on segment-specific catchment area falling inside a polygon. See more details below.

If line-based weights (equal, length, strahler, user specified vector) are used, the river network is split at grid cell boundaries before determining polygon-segment relationship.

"area" weights can be used when catchment-based weighting is desired. If no further data is supplied, the function computes a Voronoi diagram from the river network (see river_voronoi for details). If the basins are known, they can be supplied which allows skipping the delineation step entirely.

Area of interest is optional. If provided, the river network will be clipped using the supplied AoI.