Computes optimal estimate from multiple input discharge timeseries

Function performs timeseries optimisation of discharge estimates against observed streamflow timeseries at specific river segments.

optimise_point(HSflow, HSobs, optim_method = "CLS", train = 0.5,
  bias_correction = FALSE)

Arguments

HSflow
An HSflow object

HSobs
An HSobs object containing observations at the locations to optimise.

optim_method
Method used to optimise. Default uses constrained linear regression. See details.

train
The share of timeseries used for training period.

bias_correction
Apply (constant) bias correction. Defaults to FALSE.

Value

Returns an object of class HSoptim, which is a list of results from observation stations. Each list item contains:

Observations: Supllied observation timeseries.

Forecast_train: Forecasted timeseries at training period.

Forecast_test: Forecasted timeseries at testing period.

Forecast_train_test: Forecasted timeseries with training and testing combined

Method: Used method for forecast combination.

Forecast_weights: Vector of weights - how the input discharge timeseries were combined.

Intercept: Intercept from the combination. NA, if not applicable to the Method.

Goodness_of_fit. Goodness of fit of the forecast combination obtained using gof.

Details

Optimisation of the input timeseries against observations makes use of forecastComb package, with three options currently supported; "CLS" (constrained linear regression), "OLS" (ordinary linear regression), and "factorCLS" (constrained linear regression in factorized form). Default is constrained linear regression, which constraints the coefficients to add to unity, and all weights are between 0, and 1. CLS also forces Intercept to 0. In some cases, CLS fails to find a unique solution. In those cases, you can try the factorized CLS.

If bias correction is set to TRUE, a constant bias correction is applied to the entire timeseries so that mean error in the training period is 0.