Create segment-specific Voronoi diagram from a network.

The function creates Voronoi diagram for each segment in a directed connected river network, where the Voronoi polygons join together at network segment intersections.

river_voronoi(river, aoi, riverID = "riverID", min = 10,
  tolerance = 0.001, verbose = FALSE)

Arguments

river
An 'sf' linestring feature representing a river network.

aoi
An 'sf' polygon object. Optional.

riverID
A character string which specifies the name of the column in river containing unique river network identifiers. Defaults to "riverID".

min
A numeric value. Removes segments which are shorter than the value in meters, in order to remove unnecessarily short river segments. Defaults to 10 meters.

tolerance
Radius of the buffer used to take difference. Defaults to 0.001 (in map units).

verbose
Whether or not print progress indicators.

Value

Returns an 'sf' polygon object, with a column "ID" corresponding to the river segment IDs.

Details

Creating the segment Voronoi polygons is done in the following steps:

Extract segment end nodes.

Create a buffer of the end nodes, and take difference between the river network and the buffer layer.

Extract all nodes of the river network, and create Voronoi polygons for all nodes.

Dissolve the individual polygons using river segment ID.

Clip the polygons to the area of interest.

Process erroneous polygons if any.

Thus, the accuracy of the final segment Voronoi diagram is depending on the density of nodes in the river network. Consider densifying geometry e.g. with sf::st_segmentize function for higher accuracy.